--- views/index.jade
@@ -1,8 +1,0 @@
-extends layout
-
-block content
-  h1= title
-  p Welcome to #{Account}
-  p signin : localhost:3000/account/signin
-  p signup : localhost:3000/account/signup
-
